Move ArrayMap and flatArray
